- database actually works now - changed ids from ints to longs

+++ app/src/main/java/com/g2/musiq/TrackInfo.java
@@ -4,14 +4,14 @@ import android.media.Image;
 
 public class TrackInfo {
     public int requestId;
-    public int mediaId;
+    public long mediaId;
     public String mediaLocator;
     public String mediaSourceType;
     public String trackName;
     public float trackLengthSeconds;
     public Artist artist;
     public Album album;
-    public int genreId;
+    public long genreId;
     public int trackNumber;
     public Image image;
     public String lyrics;
